Initial commit for Scheduler object
Refs #9000
